requirements.txt
google-cloud-spanner==1.19.0
futures==3.3.0; python_version < "3"
python-dotenv
google-cloud
